Refactoring hyperspy exporter
